requirements.txt
asgiref==3.2.10
astroid==2.4.2
Babel==2.8.0
beautifulsoup4==4.8.2
bootstrap-admin==0.4.4
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-autoslug==1.9.8
django-braces==1.14.0
django-cogwheels==0.3
django-countries==6.1.3
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-extensions==3.0.3
django-extra-views==0.13.0
django-haystack==3.0b2
django-model-utils==4.0.0
django-modelcluster==5.0.2
django-oscar==2.1
django-phonenumber-field==3.0.1
django-rated-reviews==0.6.0
django-star-ratings==0.9.1
django-tables2==2.2.1
django-taggit==1.3.0
django-tailwind==0.9.0
django-treebeard==4.3.1
django-uuslug==1.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
draftjs-exporter==2.1.7
factory-boy==2.12.0
Faker==4.1.2
html5lib==1.1
idna==2.10
isort==4.3.21
l18n==2018.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
phonenumbers==8.12.8
Pillow==7.2.0
pkg-resources==0.0.0
purl==1.5
pycparser==2.20
pylint==2.5.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
swapper==1.1.2.post1
text-unidecode==1.3
toml==0.10.1
Unidecode==1.1.1
urllib3==1.25.10
wagtail==2.9.3
wagtailmenus==3.0.2
webencodings==0.5.1
Werkzeug==1.0.1
Willow==1.3
wrapt==1.12.1
XlsxWriter==1.2.9
